requirements.txt
ollama
pyttsx3
pydub
moviepy==1.0.3
Pillow>=12.2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
selenium
webdriver-manager
tqdm
rich
# ── Video Clipper (option 10) dependencies ──────────────────────
yt-dlp>=2024.1
ffmpeg-python>=0.2.0
librosa>=0.10.1
soundfile>=0.12.1
faster-whisper>=1.0.3
torch>=2.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
opencv-python>=4.9.0
mediapipe>=0.10.9
numpy>=1.24
scipy>=1.10
python-dateutil>=2.8
